@@ -1,12 +1,19 @@
-import {useState} from "react";
+import {useState, flushSync} from "react";
 
 function TestComponent() {
     const [count, setCount] = useState(0);
     const [flag, setFlag] = useState(false);
 
     function handleClick() {
-        setCount(c => c + 1);
-        setFlag(f => !f);
+        setTimeout(() => {
+            flushSync(() => {
+                setCount(c => c + 1);
+            })
+
+            flushSync(() => {
+                setFlag(f => !f);
+            })
+        }, 100)
     }
     console.log('render');
 
